requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.3.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
PyJWT==2.8.0
pymongo==4.5.0
python-decouple==3.8
python-dotenv==0.21.1
Werkzeug==2.2.3
zipp==3.15.0
